# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPIS API DOCUMENTATION",
        "description": "This is the Project Progres Information System API documentation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/parentUploads": {
            "get": {
                "tags": [
                    "parentUploads"
                ],
                "summary": "Get a list of parentUploads",
                "operationId": "0b18a7cba4586f06c6f7be72d896f077",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "parent_upload_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "uuid": {
                                                        "type": "string",
                                                        "example": "efdbd310-5cb7-4e94-b2dd-010185ddac95"
                                                    },
                                                    "parent_upload_name": {
                                                        "type": "string"
                                                    },
                                                    "isSelected": {
                                                        "type": "boolean"
                                                    },
                                                    "created_by": {
                                                        "type": "integer"
                                                    },
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "middle_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "parentUploads"
                ],
                "summary": "Store a new parentUploads",
                "operationId": "ea99677cf2a26904e4b79f591e578f41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_upload_name": {
                                        "type": "string"
                                    },
                                    "education_level_id": {
                                        "type": "string"
                                    },
                                    "upload_type_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}